only phi is double for now

+++ Core/include/Acts/Seeding/detail/CylindricalSpacePointGrid.hpp
@@ -37,25 +37,25 @@ using CylindricalBinnedGroupIterator = Acts::BinnedGroupIterator<
 
 struct CylindricalSpacePointGridConfig {
   // minimum pT to be found by seedFinder
-  double minPt = 0 * Acts::UnitConstants::MeV;
+  float minPt = 0 * Acts::UnitConstants::MeV;
   // maximum extension of sensitive detector layer relevant for seeding as
   // distance from x=y=0 (i.e. in r)
-  double rMax = 320 * Acts::UnitConstants::mm;
+  float rMax = 320 * Acts::UnitConstants::mm;
   // maximum extension of sensitive detector layer relevant for seeding as
   // distance from x=y=0 (i.e. in r)
-  double rMin = 0 * Acts::UnitConstants::mm;
+  float rMin = 0 * Acts::UnitConstants::mm;
   // maximum extension of sensitive detector layer relevant for seeding in
   // positive direction in z
-  double zMax = 0 * Acts::UnitConstants::mm;
+  float zMax = 0 * Acts::UnitConstants::mm;
   // maximum extension of sensitive detector layer relevant for seeding in
   // negative direction in z
-  double zMin = 0 * Acts::UnitConstants::mm;
+  float zMin = 0 * Acts::UnitConstants::mm;
   // maximum distance in r from middle space point to bottom or top spacepoint
-  double deltaRMax = 0 * Acts::UnitConstants::mm;
+  float deltaRMax = 0 * Acts::UnitConstants::mm;
   // maximum forward direction expressed as cot(theta)
-  double cotThetaMax = 0;
+  float cotThetaMax = 0;
   // maximum impact parameter in mm
-  double impactMax = 0 * Acts::UnitConstants::mm;
+  float impactMax = 0 * Acts::UnitConstants::mm;
   // minimum phi value for phiAxis construction
   double phiMin = -std::numbers::pi;
   // maximum phi value for phiAxis construction
@@ -118,7 +118,7 @@ struct CylindricalSpacePointGridConfig {
 
 struct CylindricalSpacePointGridOptions {
   // magnetic field
-  double bFieldInZ = 0.;
+  float bFieldInZ = 0.;
   bool isInInternalUnits = false;
   CylindricalSpacePointGridOptions toInternalUnits() const {
     if (isInInternalUnits) {
